Move event sorting logic into a discrete module
