+++ src/parser/core/modules/index.js
@@ -13,6 +13,7 @@ import Death from './Death'
 import Downtime from './Downtime'
 import Enemies from './Enemies'
 import {EntityStatuses} from './EntityStatuses'
+import {EventOrder} from './EventOrder'
 import GlobalCooldown from './GlobalCooldown'
 import HitType from './HitType'
 import Invulnerability from './Invulnerability'
@@ -44,6 +45,7 @@ export default [
 	Downtime,
 	Enemies,
 	EntityStatuses,
+	EventOrder,
 	GlobalCooldown,
 	HitType,
 	Invulnerability,
